+++ src/components/nav.tsx
@@ -1,6 +1,42 @@
 import { Link } from "react-router-dom";
+import { useEffect, useState } from "react";
 
 export default function Navbar() {
+    const [isLoggedIn, setIsLoggedIn] = useState(false)
+
+    const profileUrl = "/profile"
+    const loginUrl = "/login"
+
+    useEffect(() => {
+        const token = localStorage.getItem("Authorization")
+
+        if (!token) {
+            setIsLoggedIn(false)
+            return
+        }
+
+        async function validateToken() {
+            try {
+                const response = await fetch('http://localhost:8000/v1/verify', {
+                    method: 'POST',
+                    headers: { 'Authorization': `Bearer ${token}` },
+                });
+
+                // Could use this later for maybe profile picture URL etc
+                // const result = await response.json();
+                if (response.ok) {
+                    setIsLoggedIn(true);
+                } else {
+                    setIsLoggedIn(false);
+                }
+            } catch {
+                setIsLoggedIn(false);
+            }
+        }
+
+        validateToken();
+    }, [])
+
     return (
         <nav className="bg-zinc-900 text-white p-4 shadow-md font-outfit">
             <div className="container mx-auto flex justify-between items-center">
@@ -11,7 +47,7 @@ export default function Navbar() {
                 >
                     New Wave <span className="text-amber-100">Fashion</span>
                 </Link>
-                
+
                 {/* Links */}
                 <ul className="flex space-x-6">
                     <li>
@@ -48,10 +84,10 @@ export default function Navbar() {
                     </li>
                     <li>
                         <Link
-                            to="/login"
+                            to={isLoggedIn ? profileUrl : loginUrl}
                             className="hover:text-shadow-white transition-all duration-300"
                         >
-                            Login
+                            {isLoggedIn ? "Profile" : "Login"}
                         </Link>
                     </li>
                 </ul>